requirements.txt
requests
pandas
matplotlib
xmltodict
python-dotenv[cli]
pylint